python/functions/hello-gcs/requirements.txt
flask==2.1.0
functions-framework==3.2.1
google-events==0.4.0
